feat: Refactor zBrick and zCallableAugmentation classes
This commit refactors the `zBrick` and `zCallableAugmentation` classes to improve code organization and separation of concerns. The `_post_init` method in the `zBrick` class has been commented out as it is no longer needed. Additionally, the `zWsgiApplication` class has been updated to utilize the new class hierarchy and remove the `_post_init` method.